print all

+++ 0x10-variadic_functions/3-print_all.c
@@ -30,7 +30,7 @@ void print_all(const char * const format, ...)
 				break;
 			case 's':
 				str = va_arg(args, char *);
-				if (str == NULL)
+				if ((int)(str == NULL))
 				{
 					str = "(nil)";
 					printf("%s%s", sep, str);
